# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Vyapaar AI API Specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analytics/dashboard:

    get:
      operationId: getAnalyticsDashboard
      tags: [analytics]
      summary: Get Analytics Dashboard Stats
      description: Retrieves core KPIs, sales, CRM pipeline conversion numbers, and active comparisons.
      parameters:
        - name: orgId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analytics dashboard summaries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsSummary"
  /analytics/forecast/inventory:
    get:
      operationId: getInventoryForecast
      tags: [analytics]
      summary: Predict Product Stockout Risks
      description: Flags products likely to stock out in the next N days based on sales velocity.
      parameters:
        - name: orgId
          in: query
          required: true
          schema:
            type: integer
        - name: days
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Stockout risk predictions list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    productId:
                      type: integer
                    name:
                      type: string
                    currentStock:
                      type: integer
                    velocity:
                      type: number
                    daysRemaining:
                      type: integer
                    riskLevel:
                      type: string
                  required: [productId, name, currentStock, velocity, daysRemaining, riskLevel]
  /analytics/forecast/revenue:
    get:
      operationId: getRevenueForecast
      tags: [analytics]
      summary: Forecast Next Month's Revenue
      description: Linear trend revenue forecasting projection with merchant warning caveats.
      parameters:
        - name: orgId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Monthly revenue forecasting details
          content:
            application/json:
              schema:
                type: object
                properties:
                  projectedRevenue:
                    type: integer
                  confidenceLevel:
                    type: string
                  caveat:
                    type: string
                  projectedMonth:
                    type: string
                  historicalTrend:
                    type: array
                    items:
                      type: object
                      properties:
                        month:
                          type: string
                        revenue:
                          type: integer
                      required: [month, revenue]
                required: [projectedRevenue, confidenceLevel, caveat, projectedMonth, historicalTrend]
...
components:
  schemas:
...
    AnalyticsSummary:
      type: object
      properties:
        revenueToday:
          type: integer
        revenueMonthly:
          type: integer
        ordersTotal:
          type: integer
        customersCount:
          type: integer
        salesHistory:
          type: array
          items:
            type: object
            properties:
              label:
                type: string
              val:
                type: integer
        leadsHistory:
          type: array
          items:
            type: object
            properties:
              status:
                type: string
              count:
                type: integer
      required: [revenueToday, revenueMonthly, ordersTotal, customersCount]